move GridItem and Chart States to sharedComponents
This moves:
- GridItem
- ChartStateLoading
- ChartStateNoData

To @bufferapp/analyze-shared-components
