+++ packages/average-table/package.json
@@ -23,6 +23,7 @@
     "@bufferapp/async-data-fetch": "0.5.25",
     "@bufferapp/analyze-date-picker": "0.0.1",
     "@bufferapp/analyze-profile-selector": "0.0.1",
+    "@bufferapp/analyze-shared-components": "0.0.1",
     "@bufferapp/summary-table": "0.0.1"
   },
   "devDependencies": {
