+++ packages/shared-components/GridItem/components/TruncatedNumber/index.jsx
