+++ packages/shared-components/index.js
@@ -1,3 +1,6 @@
 export PeriodPhrase from './PeriodPhrase';
 export SocialIcon from './SocialIcon';
 export ProfileBadge from './ProfileBadge';
+export ChartStateNoData from './ChartStateNoData';
+export ChartStateLoading from './ChartStateLoading';
+export GridItem from './GridItem';
