+++ packages/summary-table/index.js
@@ -12,7 +12,3 @@ export default connect(
 
 export reducer, { actions, actionTypes } from './reducer';
 export middleware from './middleware';
-
-export GridItem from './components/GridItem';
-export Loading from './components/Loading';
-export NoData from './components/NoData';
